% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_docs.R
\docType{data}
\name{Artinskian}
\alias{Artinskian}
\title{Artinskian}
\format{shapefile}
\source{
http://www.gplates.org/
}
\description{
A shapefile with the location of the continents at that temporal interval.
}
\keyword{datasets}
