% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInput.R
\name{checkDataRank}
\alias{checkDataRank}
\title{checkDataRank}
\usage{
checkDataRank(data, rank)
}
\arguments{
\item{data}{a data frame with fossil occurences}

\item{rank}{a string defning the rank of interest}
}
\value{
boolean of correct columns are inside the data frame
}
\description{
checks if the column corresponding to the rank is inside the data frame
}
\examples{
\dontrun{
data <- pm_getdata(base_name="Canis", interval="Quaternary")
rank <- "species"
checkDataRank(data, rank)
} 
}
