% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInput.R
\name{checkFun}
\alias{checkFun}
\title{checkFun}
\usage{
checkFun(fun, fct)
}
\arguments{
\item{fun}{a string defining the function}

\item{fct}{the fct that is calling this check. 
Either pm_latdiv or pm_divraster}
}
\value{
boolean if function is allowed or not
}
\description{
checks if chosen function is allowed for this function
}
\examples{
\dontrun{
fun1 <- mean
fct1 <- "pm_latdiv"
checkFun(fun1, fct1)

fun2 <- "count"
fct2 <- "pm_divraster"
checkFun(fun2, fct2)
}
}
