% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInput.R
\name{checkLatLng}
\alias{checkLatLng}
\title{checkLatLng}
\usage{
checkLatLng(data)
}
\arguments{
\item{data}{a data frame with fossil occurences}
}
\value{
boolean if paleolat and paleolng are columns of the data frame
}
\description{
checks if the input data frame has 
paleolatitude and paleolongitude as columns
}
\examples{
\dontrun{
data <- pm_getdata(base_name="Canis", interval="Quaternary")
checkLatLng(data)
}
}
