% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInput.R
\name{checkRank}
\alias{checkRank}
\title{checkRank}
\usage{
checkRank(rank)
}
\arguments{
\item{rank}{string with the rank of interest}
}
\value{
boolean if the rank is possible to be chosen
}
\description{
checks if the chosen rank is correct
}
\examples{
\dontrun{
rank <- "species"
checkRank(rank)
}
}
