% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInput.R
\name{checkShape}
\alias{checkShape}
\title{checkShape}
\usage{
checkShape(shape)
}
\arguments{
\item{shape}{a SpatialPolygonsDataFrame}
}
\value{
a boolean if the shape is a SpatialPolygonsDataFrame
}
\description{
checks if the input shape is a SpatialPolygonsDataFrame
}
\examples{
\dontrun{
shape <- pm_getmap(interval="Quaternary", model="GPlates")
checkShape(shape)
}
}
