% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_docs.R
\name{data}
\alias{data}
\title{Fossil dataset for tests
A dataframe with the fossil occurences}
\format{data.frame}
\description{
Fossil dataset for tests
A dataframe with the fossil occurences
}
\keyword{data}
