% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{formatdata}
\alias{formatdata}
\title{formatdata}
\usage{
formatdata(data, db = "pbdb")
}
\arguments{
\item{data}{data.frame. Fossil occurrences data from the Paleoiology Database.}

\item{db}{character. Name of the database where the data is from. Only db = "pbdb" possible.}
}
\value{
data.frame
}
\description{
Changes and adds columns to your fossil occurrences data.frame from the Paleobiology Database to the names needed for the functions in this package.
It also calculates the average age of the fossil occurrences from the data.frame columns early_age and late_age.
}
\examples{
\dontrun{

library(mapast)

data  <-  base::data.frame(paleobioDB::pbdb_occurrences(limit = 100, base_name = "Canis", 
  min_ma = 0, max_ma = 10, 
  show = c("coords", "phylo"), 
  vocab = "pbdb"))
  
fdata <- formatdata(data, db = "pbdb")

}
}
