% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{madata}
\alias{madata}
\title{madata}
\usage{
madata(minma, maxma, base_name, limit="all", paleocoords=F, 
recontime=NULL, model='SETON2012')
}
\arguments{
\item{minma}{numeric. Minimal age in Ma.}

\item{maxma}{numeric. Maximal age in Ma.}

\item{base_name}{character. The name of the taxon of interest. (e.g. "Canis" or "reptilia")}

\item{limit}{integer. Defining the max. number of occurrences to be downloaded from paleobioDB. By default limit="all"}

\item{paleocoords}{boolean. Defines if the user wants to get the paleocoordinates.}

\item{recontime}{numeric. Time of reconstruction in ma. If recontime is null the recnstruction time will be 
the midpoint of the interval the fossil occurred.}

\item{model}{character. Defining the model the map should be created with. 'SETON2012' (default), 
'MULLER2016', 'GOLONKA', 'PALEOMAP' or 'MATTHEWS2016'.}
}
\value{
Returns a data frame with fossil occurrences
}
\description{
Uses the paleobioDB R package to extract data used in for other functions of this package
from the Paleobiology Database.
}
\examples{
\dontrun{

#get data without paleocoordinates
occ <- madata(minma=0, maxma=10,base_name="Canis", limit=100)

#get data with paleocoordinates taking the midpoint of the occurrence time
occma <- madata(minma=0, maxma=10,base_name="Canis", 
                    limit=100, paleocoords=TRUE, model='SETON2012')

#get data with paleocoordinates for a chosen time in ma
occmatime <- madata(minma=0, maxma=10,base_name="Canis", limit=100, 
                         paleocoords=TRUE, recontime=5,  model='SETON2012')
                    
                    
}
}
