% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapast_package.R
\docType{package}
\name{mapast R-package}
\alias{mapast R-package}
\alias{mapast-package}
\title{mapast: MAp the PAST.}
\description{
We have developed mapast (MAp the PAST), an R-package designed to make paleontolgical maps available to 
visualise the fossil occurrences from the Paleobiology Database onto their corresponding maps and analyse them.
We programmed functions for getting the SpatialPolygonsDaraFrame of the paleontological maps, 
functions to get a data.frame of fossil occurrences from the Paleobilogy Database, analysing the diversity of the past 
and plotting this data on the maps.
}
\details{
\tabular{ll}{
Package: \tab mapast\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2017-12-12\cr
License: \tab GPL-2\cr
}
}
\seealso{
GPlates
\url{http://www.gplates.org/}

Golonka & Smith Global Paleoshorelines
\url{https://github.com/chhei/Heine_AJES_15_GlobalPaleoshorelines}

PaleobioDB
\url{http://paleobiodb.org}
}
\author{
Sonja Rothkugel \email{rothkugelsonja@aol.de}

Sara Varela \email{svarela@paleobiogeography.org}
}
\keyword{package}
