% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mapocc}
\alias{mapocc}
\title{mapocc}
\usage{
mapocc(data, model = "SETON2012",
                   rank = "genus", map = NULL,
                   res = 1, save.as = NULL,
                   colland = "#66666660",
                   colsea = "#00509010", col.grid = mycols(100), do.plot = TRUE, ...)
}
\arguments{
\item{data}{data.frame. Fossil occurrences data.}

\item{model}{character. Defining the model the map should be created with. "SETON2012" (default), 
"MULLER2016", "GOLONKA", "PALEOMAP" or "MATTHEWS2016".}

\item{rank}{character. Defining the taxonomic rank of interest. 
"species", "genus", "family", "order", "class" or "phylum". By default rank = "genus".}

\item{map}{(list of) SpatialPolygonDataFrames. Containing map(s) which can be created by getmap.}

\item{res}{numeric. Defining the spatial resolution. By default res = 1.}

\item{save.as}{character. Defines the format the plots should be saved. "tiff", "pdf", "jpeg" or "png".}

\item{colland}{character. Defines the color of the land masses. By default colland = "#66666660".}

\item{colsea}{character. Defines the color of the sea. By default colsea = "#00509010".}

\item{col.grid}{character. Defines the color of the raster. By default col.grid = mycols(100),}

\item{do.plot}{logical. Defines if a plot is created or not. By default do.plot = TRUE.}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to
plot, such as main = "my own title" or main.col = "red".}
}
\value{
RasterLayer
}
\description{
Creates a RasterLayer, containing the number of occurrences per cell and a plot of the fossil 
occurrences by the taxonomic rank per cell (a proxy for the sampling effort).
}
\examples{
\dontrun{

library(mapast)

# get data and preprocess it
data  <-  base::data.frame(paleobioDB::pbdb_occurrences(base_name = "Canis", 
                                                        min_ma = 0, max_ma = 10, 
                                                        show = c("coords", "phylo"), 
                                                        vocab = "pbdb", limit = 100))
df <- formatdata(data = data)
df_auto <- paleocoords(data = df, time = "automatic")
                                                        
# create a plot with fossils on the paleogeographical map
occras <- mapocc(data = df_auto, model = "SETON2012", rank = "species")

# save the maps before so the function does not need to load them
maps <- getmap(ma = 2.5, model = "SETON2012", do.plot = FALSE)
mapocc(data = df_auto, model = "SETON2012", rank = "species", map = maps)

# save maps as pdf
mapocc(data = df_auto, model = "SETON2012", rank = "species", map = maps, save.as = "pdf")

}
}
