% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{modelmap}
\alias{modelmap}
\title{modelmap}
\usage{
modelmap(ma, model = 'SETON2012', show.plates = FALSE, colland, colsea, do.plot, ...)
}
\arguments{
\item{ma}{numeric. Age in Ma.}

\item{model}{character. Defining the model the map should be created with. 'SETON2012' (default), 
'MULLER2016', 'GOLONKA', 'PALEOMAP' or 'MATTHEWS2016'.}

\item{show.plates}{boolean. Defines if the user wants to get the continental plate borders. By default. show.plates = FALSE.}

\item{colland}{character. Defines the color of the land masses. By default colland = "#66666660".}

\item{colsea}{character. Defines the color of the sea. By default colsea = "#00509010".}

\item{do.plot}{logical. Defines if a plot is created or not. By default do.plot = TRUE.}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to plot, such as main="my own title", main.col="red"}
}
\value{
SpatialPolygonsDataFrame
}
\description{
Gets a map of a specific age from a model specified by the user.
Available models and ages at https://github.com/GPlates/gplates_web_service_doc/wiki/Reconstruction-Models .
}
\examples{
\dontrun{

#with continental plates
map <- modelmap(100, model = 'SETON2012', show.plates = T)
coastlines <- map[[1]]
plates <- map[[2]]

#without continental plates
coastlines <- modelmap(100, model = 'SETON2012')

#save map as tiff image
tiff("getmap-GPlates_Cretaceous.tiff", 
      height = 10, width = 17, units = "cm", res = 300)
modelmap(100, model = 'SETON2012', show.plates = T)
dev.off()

}
}
