% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{paleocoords}
\alias{paleocoords}
\title{paleocoords}
\usage{
paleocoords(occ, recontime=NULL, model='SETON2012')
}
\arguments{
\item{occ}{data.frame. Fossil data, which should at least have the columns lng and lat (if recontime=NULL also early_age and late_age).}

\item{recontime}{numeric. Time of reconstruction in ma. If recontime is null the recnstruction time will be 
the midpoint of the interval the fossil occurred.}

\item{model}{character. Defining the model the map should be created with. 'SETON2012' (default), 
'MULLER2016', 'GOLONKA', 'PALEOMAP' or 'MATTHEWS2016'.}
}
\value{
data.frame
}
\description{
Descr.
}
\examples{
\dontrun{

occ <- madata(minma=0, maxma=10,base_name="Canis", limit=100)

#reconstruct paleocoordinates with midpoint of appearance time
occ_ma <- paleoocoords(occ, recontime=NULL, model='SETON2012')

#reconstruct paleocoordinates with specific time
occ_matime <- paleoocoords(occ, recontime=5, model='SETON2012')

                    
}
}
