% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasta_package.R
\docType{package}
\name{pasta}
\alias{pasta}
\alias{pasta-package}
\alias{pasta-package}
\title{pasta (PAST mAps):  An R-package for getting and using paleogeographical maps.}
\description{
We have developed pasta (PAST mAps), an R-package designed to make paleontolgical maps available to
plot the fossil occurrences from the Paleobiology Database onto paleogeographical maps.
}
\details{
We programmed functions for getting the SpatialPolygonsDaraFrame of the paleontological maps, 
functions to get a data.frame of fossil occurrences from the Paleobilogy Database, analysing the diversity of the past 
and plotting this data on the maps.

\tabular{ll}{
Package: \tab pasta\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2017-12-12\cr
License: \tab GPL-2\cr
}
}
\seealso{
GPlates
\url{http://www.gplates.org/}

Golonka & Smith Global Paleoshorelines
\url{https://github.com/chhei/Heine_AJES_15_GlobalPaleoshorelines}

PaleobioDB
\url{http://paleobiodb.org}
}
\author{
Sonja Rothkugel \email{rothkugelsonja@aol.de}

Sara Varela \email{svarela@paleobiogeography.org}
}
\keyword{package}
