% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pastplot}
\alias{pastplot}
\title{pastplot}
\usage{
pastplot(interval, model, data, colland = "#66666660",
                colsea = "#00509010",colpoints = "#65432190", 
                pch = 16, cex = 1, ...)
}
\arguments{
\item{interval}{character. Temporal paleoeographical interval of interest. (e.g. "Cretaceous" for GPlates or "112.0" for Smith)}

\item{model}{character. Defining the model the map was created with. "GPlates", "Smith" or "Golonka".}

\item{data}{data.frame. Fossil occurrences data. Can be created with getdata(interval, base_name).}

\item{colland}{character. Defines the color of the land masses. By default colland = "#66666660".}

\item{colsea}{character. Defines the color of the sea. By default colsea = "#00509010".}

\item{colpoints}{character. Defines the color of the occurrence-points. By default colpoints = "#65432190".}

\item{pch}{numeric. Point symbol for plotting the occurences. By default pch = 16 (filled circle).}

\item{cex}{numeric. Size of the points. By default cex = 1.}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to plot, such as main="my own title", main.col="red"}
}
\value{
Plot
}
\description{
Plots your fossil occurrence data from the paleobioDB onto the map of the selected time interval.
}
\examples{
\dontrun{

library(pasta)

#get data
data  <-  getdata (interval = "Cretaceous", base_name = "Mammalia")
#create a plot with fossils on the paleogeographical map
pastplot(interval = "Cretaceous", model = "GPlates", data)

#save plot as pdf file
pdf("pastplot-GPlates_Cretaceous-Mammalia.pdf")
pastplot(interval = "Cretaceous", model = "GPlates", data)
dev.off()

#save plot as tiff image
tiff("pastplot-GPlates_Cretaceous-Mammalia.tiff", 
      height = 10.5, width = 17, units = "cm", res = 300)
pastplot(interval = "Cretaceous", model = "GPlates", data)
dev.off()

}
}
