% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_checkAge}
\alias{pm_checkAge}
\title{pm_checkAge}
\usage{
pm_checkAge(age)
}
\arguments{
\item{age}{character. defining the age of interest.}
}
\value{
a data.frame which maps are available
}
\description{
creates a data frame with all maps that contain the age requested
}
\examples{
\dontrun{
pm_checkAge(age="112")
}
}
