% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_corlatrich}
\alias{pm_corlatrich}
\title{pm_corlatrich}
\usage{
pm_corlatrich (ngl_data)
}
\arguments{
\item{ngl_data}{a data frame with number of genus per locality
Can be created with pm_getdata(interval, base_name)}
}
\value{
data frame with shannon corrected richness of rank
}
\description{
calculates the shannon corrected latitudinal richness of the genus
}
\examples{
\dontrun{
data<- pm_getdata (base_name="Canis", interval="Quaternary")
ngl_data <- pm_ngl(data)
corlatrich <- pm_corlatrich (ngl_data)
show(corlatrich)
}
}

