% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_corrichraster}
\alias{pm_corrichraster}
\title{pm_corrichraster}
\usage{
pm_corrichraster (shape, ngl_data, res, colsea, colland, colborder)
}
\arguments{
\item{shape}{file from the time interval of interest.
Can be created with get_paleomap}

\item{ngl_data}{a data frame with number of genus per locality}

\item{res}{resolution of the raster/ size of the grid cell}

\item{colsea}{color of the ocean}

\item{colland}{color of the land masses}

\item{colborder}{color of the landmass borders}
}
\value{
plot with map of the time intervall, the fossil occurences and the
raster file. And the raster file itself
}
\description{
creates a raster of species richness
and makes a plot of the map and raster
}
\examples{
\dontrun{
shape<- pm_getdata (base_name="Canis", interval="Quaternary")
data<- pm_getdata (base_name="Canis", interval="Quaternary")
ngl_data <- pm_ngl(data)
myraster <- pm_corrichraster (shape, ngl_data)
plot(myraster)
}
}

