% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_divraster_cell}
\alias{pm_divraster_cell}
\title{pm_divraster_cell}
\usage{
pm_divraster_cell  (shape, occ_df_cell, res=10,
                           colland="#66666680", colsea="#00509010", ...)
}
\arguments{
\item{shape}{SpatialPolygonsDataFrame object containing a map.}

\item{occ_df_cell}{data.frame with fossil occurrences. Can be created with pm_occ_cell (data)}

\item{res}{numeric. Defining the spatial resolution. Default res=10.}

\item{colland}{define the color of the land masses. By default colland = "#66666660".}

\item{colsea}{define the color of the sea. By default colsea = "#00509010".}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to
plot, such as main="my own title", main.col="red".}
}
\value{
plot with map of the time intervall, the fossil occurences and the 
raster file. And the raster file itself
}
\description{
calculates the Shannon diversity per cell 
(taking into account relative abundances of all the fossil records 
whithin the cell)
}
\examples{
\dontrun{
shape<- pm_getmap(interval="Quaternary", model="GPlates", do.plot=FALSE)
data<- pm_getdata (base_name="Canis", interval="Quaternary")
occ_df_cell <- pm_occ_cell (data, rank="species")
div_cell <- pm_divraster_cell (shape, occ_df_cell, res=10)
## cells with diversity values = 0 (e.g., 1 species) are discarded.
}
}
