% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_latdiv}
\alias{pm_latdiv}
\title{pm_latdiv}
\usage{
pm_latdiv (shape, occ_df, res=1, fun= max,
                  colland="#66666680", colsea="#00509010",  
                  colpoints="#65432190",
                  div.col="#654321", pch=21, do.plot=TRUE, ...)
}
\arguments{
\item{shape}{SpatialPolygonsDataFrame object containing a map.}

\item{occ_df}{data.frame with fossil occurrences. Can be created with pm_occ(data)}

\item{res}{numeric. Defining the spatial resolution. By default res=1.}

\item{fun}{function or character. To determine what values to assign to cells that are covered by multiple spatial features. 
You can use functions such as min, max, or mean, or the character value: 'count'.}

\item{colland}{define the color of the land masses. By default colland = "#66666660".}

\item{colsea}{define the color of the sea. By default colsea = "#00509010".}

\item{colpoints}{define the color of the occurrence-points. By default colpoints="#65432190".}

\item{div.col}{define the color od zje diversity curve. By default div.col="#654321".}

\item{pch}{point symbol for plotting the occurences. By default pch=21.}

\item{do.plot}{logical. Defines if a plot is created or not. By default do.plot=TRUE.}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to plot, such as main="my own title", main.col="red"}
}
\value{
Returns adata frame with shannon diversity,
a plot of the corresponding time map with the occurrences and their
latitudinal diversity
}
\description{
Calculates the Shannon diversity along the latitudinal gradient based on 
the individual values of diverstiy the fossil localities of those latitudes.
The function returns the mean or max values of diversity of the sampled 
localities along the latitudinal gradient.
}
\examples{
\dontrun{

shape<- getmap(interval="Quaternary", model="GPlates", do.plot=FALSE)
data<- getdata (base_name="Canis", interval="Quaternary")
occ_df <- pm_occ (data)
pm_latdiv (shape, occ_df, fun=mean)
pm_latdiv (shape, occ_df, fun=max)

#save as pdf file
pdf("pm_latrich-GPlates_Quaternary-Canis.pdf")
pm_latdiv (shape, occ_df, fun=mean)
dev.off()
#save as tiff image
tiff("pm_latrich-GPlates_Quaternary-Canis.tiff", 
      height = 9, width = 17.5, units = 'cm', res=300)
pm_latdiv (shape, occ_df, fun=mean)
dev.off()

}
}
