% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_latrich}
\alias{pm_latrich}
\title{pm_latrich}
\usage{
pm_latrich (shape, data, res=10, rank="species",
                   colsea="#00509010", colland="#66666680",
                   colpoints="#FFC12530",colpointborder="black", 
                   rich.col="goldenrod1", pch=21, magn=1, ...)
}
\arguments{
\item{shape}{a shape file of the corresponding time map}

\item{data}{a data frame with fossil occurrences 
Can be created with pm_getdata(interval, base_name)}

\item{res}{resolution in of the segmentation of the latitude. By default res=1.}

\item{rank}{character, taxonomic rank: "species", "genus", "family", "order". 
By default rank="species"}

\item{colsea}{defines the color of the sea}

\item{colland}{defines the color f the landmasses}

\item{colpoints}{defines the colo of the points for the occurrences}

\item{colpointborder}{defines color of the border of the occurrence points}

\item{rich.col}{color for richnness graph}

\item{pch}{point symbol for plotting the occurences. By default pch=16 (filled circle)}

\item{magn}{numeric. index to magnify the plot of the latitudinal richness 
(use when richness values are very low to see the latitudinal pattern more easily)}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to plot, such as main="my own title", main.col="red"}
}
\value{
data frame with richness of rank and a plot of the continental masses 
with the occurrences and the latitudinal richness
}
\description{
calculates latitudinal diversity of taxa (species, genera, families, orders)
}
\examples{
\dontrun{
shape<- pm_getmap(interval="Quaternary", model="GPlates", do.plot=FALSE)
data<- pm_getdata (base_name="Canis", interval="Quaternary")
pm_latrich (shape, data, rank="species", res=10)
}
}
