% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_ngl}
\alias{pm_ngl}
\title{pm_ngl}
\usage{
pm_ngl (data)
}
\arguments{
\item{data}{a data frame with fossil occurrences
Can be created with pm_getdata(interval, base_name)}
}
\value{
data frame with number of genus per locality
}
\description{
calculates the number of genus per locality
}
\examples{
\dontrun{
data<- pm_getdata (base_name="Canis", interval="Quaternary")
my_ngl <- pm_ngl (data)
show(my_ngl)
}
}

