% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_occ}
\alias{pm_occ}
\title{pm_occ}
\usage{
pm_occ(data, rank = "genus", pa=FALSE)
}
\arguments{
\item{data}{data.frame with fossil occurrences. Can be created with 
getdata(interval, base_name)}

\item{rank}{character. Defining the taxonomic rank of interest. 
"species", "genus", "family", "order", "class" or "phylum". By default rank = "genus"}

\item{pa}{boolean. Defines if the user wants presence absence or counted data. By default pa=FALSE.}
}
\value{
Returns a matrix with number of species, genera, families or orders 
per locality.
}
\description{
Generates a diversity matrix, with the number occurrences of each species, 
genus, family or order per locality.
}
\examples{
\dontrun{

data <- getdata(base_name = "Canis", interval = "Quaternary")
result <- pm_occ(data, rank = "genus")

}
}
