% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_occ_cell}
\alias{pm_occ_cell}
\title{pm_occ_cell}
\usage{
pm_occ_cell(data, rank = "genus", res = 10)
}
\arguments{
\item{data}{data.frame with fossil occurrences. Can be created with 
pm_getdata(interval, base_name)}

\item{rank}{character. Defining the taxonomic rank of interest. 
"species", "genus", "family", "order", "class" or "phylum". By default rank = "genus"}

\item{res}{numeric. Defining the spatial resolution. By default res=10.}
}
\value{
data frame with number of species, genera, families or orders per locality
}
\description{
generates a diversity matrix, with the number occurrences of each species, 
genus, family or order per cell
}
\examples{
\dontrun{
data <- pm_getdata(base_name = "Canis", interval = "Quaternary")
result <- pm_occ_cell(data, rank = "genus", res = 10)
}
}
