% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_occraster}
\alias{pm_occraster}
\title{pm_occraster}
\usage{
pm_occraster (shape, data, colsea, colland, colborder)
}
\arguments{
\item{data}{a data frame which needs to have a column called paleolat
and a column called paleolng, can be created with getdata_paleomap}

\item{colsea}{color of the ocean}

\item{colland}{color of the land masses}

\item{colborder}{color of the landmass borders}

\item{the}{shape file from the time interval of interest. Can be created with get_paleomap}
}
\value{
a raster file and a plot of the time intervall, the fossil occurences and the raster file.
}
\description{
creates a raster and a plot of the fossil occurences by taxonomic rank per cell
(a proxy for the sampling effort)
}
\examples{
\dontrun{
myraster <-  pm_occraster (shape, data, colsea, colland, colborder)
}
}

