% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_plot}
\alias{pm_plot}
\title{pm_plot}
\usage{
pm_plot (interval, base_name, limit, colsea, colland, 
colpoints, cex)
}
\arguments{
\item{interval}{time interval of interest (e.g. Quaternary)}

\item{base_name}{taxa to be downloaded for the paleobioDB (e.g Canis, Canidae, Carnivora, etc.)}

\item{limit}{enables the user to set the maximum number of 
records downloaded from the paleobioDB (e.g. 1000)}

\item{colsea}{color of the ocean}

\item{colland}{color of the land masses}

\item{colpoints}{color of the occurrence points}

\item{cex}{numeric. size of the points. By default cex=3.}
}
\value{
a plot with the configuration of the continents at the selected time interval 
and the fossil occurrences
}
\description{
plots your query from the paleobioDB onto the map of the selected time interval
}
\examples{
\dontrun{
pm_plot (interval="Quaternary", base_name= "Canis")
}

}

