% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_richraster}
\alias{pm_richraster}
\title{pm_richraster}
\usage{
pm_richraster (shape, data, res, rank, colsea, colland)
}
\arguments{
\item{shape}{file from the time interval of interest. 
Can be created with get_paleomap}

\item{data}{a data frame which needs to have a column called 
paleolat and a column called paleolng, can be created with getdata_paleomap}

\item{res}{resolution of the raster/ size of the grid cell}

\item{rank}{gives the rank for the richness raster (e.g. species, genus,...)}

\item{colsea}{color of the ocean}

\item{colland}{color of the land masses}
}
\value{
plot with map of the time intervall, the fossil occurences and the 
raster file. And the raster file itself
}
\description{
creates a raster of species richness
and makes a plot of the map and raster
}
\examples{
\dontrun{
shape<- pm_getmap(interval="Quaternary") 
data<- pm_getdata (base_name="Canis", interval="Quaternary")
richness<- pm_richraster (shape, data, rank="genus")
}

}

