% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{rank_filter}
\alias{rank_filter}
\title{rank_filter}
\usage{
rank_filer (data, res, rank)
}
\arguments{
\item{data}{a data frame which needs to have a column called paleolat and a column called paleolng,
can be created with getdata_paleomap}

\item{res}{resolution of the raster file}

\item{rank}{rank of interest}
}
\value{
filtered data frame with only species
}
\description{
filters the data frame so tehre are only species left
and for each raster every species only once
}
\examples{
\dontrun{
filtered_data <- rank_filter (data, res, rank)
show(data)
}
}

