% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{rfilter}
\alias{rfilter}
\title{rfilter}
\usage{
rfiler (data, rank)
}
\arguments{
\item{data}{a data frame which needs to have a column called paleolat and a column called paleolng,
can be created with getdata_paleomap}

\item{rank}{rank of interest}
}
\value{
filtered data frame with only species
}
\description{
filters the data frame so tehre are only species left
and for each raster every species only once
}
\examples{
/dontrun{
data<- pm_getdata (base_name="Canis", interval="Quaternary")
filtered_data <- rfilter (data, rank)
show(filtered_data)
}
}

