% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{spsite}
\alias{spsite}
\title{spsite}
\usage{
spsite(data, unity, res = 1, rank = "genus", pa = FALSE)
}
\arguments{
\item{data}{data.frame. Fossil occurrences data.}

\item{unity}{character. If the user wants the occurrences per 
cell (unity = "cell") or per fossilsite (unity = "fossilsite).}

\item{res}{numeric. The spatial resolution. Only used if unity = "cell". By default res = 1.}

\item{rank}{character. The taxonomic rank of interest. 
"species", "genus", "family", "order", "class" or "phylum". By default rank = "genus"}

\item{pa}{boolean. Defines if the user wants presence/absence or counted data. By default pa = FALSE, which means the function returns counted data.}
}
\value{
data.frame
}
\description{
Generates a diversity data.frame, with the number occurrences of taxa per locality or a presence/absence data.frame.
}
\examples{
\dontrun{

library(mapast)

#get data and preprocess it
data  <-  base::data.frame(paleobioDB::pbdb_occurrences(base_name = "Canis", 
                                                        min_ma = 0, max_ma = 10, 
                                                        show = c("coords", "phylo"), 
                                                        vocab = "pbdb", limit = 100))
df <- formatdata(data = data)
df_auto <- paleocoords(data = df, time = "automatic")
                                                        
#create a plot with fossils on the paleogeographical map
sp_fossilsite <- spsite(data = df_auto, unity = "fossilsite", rank = "species", res = 10)
sp_cell <- spsite(data = df_auto, unity = "cell", rank = "species", res = 10)


}
}
