% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_docs.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{testdata}
\format{shapefile, data.frame}
\source{
http://www.gplates.org/
}
\description{
A shapefile from the Jurassic time interval and fossil data from reptiles from the Jurassic (limit=1000)
}
\keyword{datasets}
