% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{checkage}
\alias{checkage}
\title{checkage}
\usage{
checkage(age = "all")
}
\arguments{
\item{age}{character. Defining the age of interest in mya. By default age = "all", which gives the complete list.}
}
\value{
data.frame
}
\description{
Creates a data.frame with all maps that contain the requested age.
}
\examples{
\dontrun{

library(pasta)

#get a data.frame with all maps including the age 112 mya
checkage(age = "112")

#get complete data.frame with all available ages, from all models
checkage()

}
}
