% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_help_docs.R
\docType{data}
\name{df_maps}
\alias{df_maps}
\title{df_maps}
\format{data.frame}
\description{
Dataset. A data.frame containing model, interval, fromage and toage of the maps provided by the package.
}
\keyword{datasets}
