% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{formatdata}
\alias{formatdata}
\title{formatdata}
\usage{
formatdata(data, db = "pbdb")
}
\arguments{
\item{data}{data.frame. Fossil occurrences data.}

\item{db}{character. Name of the database where the data is from.}
}
\value{
data.frame
}
\description{
Changes columns names of your data.frame from a fossil database to the names needed for the functions in this package.
It also calculates the average age if not already given by the database.
}
\examples{
\dontrun{

data  <-  base::data.frame(paleobioDB::pbdb_occurrences(limit = 100, base_name = "Canis", 
  min_ma = 0, max_ma = 10, 
  show = c("coords", "phylo"), 
  vocab = "pbdb"))
fdata <- formatdata(data, db = "pbdb")

}
}
