% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{latdivgrad}
\alias{latdivgrad}
\title{latdivgrad}
\usage{
latdivgrad (data, method, rank = "genus",
                        res = 1, map = NULL, model = "SETON2012",
                        colland = "#66666680", colsea = "#00509010", 
                        colpoints = "#65432190",
                        rich.col = "#654321", pch = 21, 
                        do.plot = TRUE, save.as = NULL,...)
}
\arguments{
\item{data}{data.frame. Fossil occurrence data.}

\item{method}{character. Defining the method of diversity measure, method = "shannon" or method = "richness".}

\item{rank}{character. Defining the taxonomic rank of interest. 
"species", "genus", "family", "order", "class" or "phylum". By default rank = "genus".}

\item{res}{numeric. Defining the spatial resolution. Default res = 1.}

\item{map}{(list of) SpatialPolygonDataFrames. Containing map(s) which can be created by getmap.}

\item{model}{character. Defining the model the map should be created with. "SETON2012" (default), 
"MULLER2016", "GOLONKA", "PALEOMAP" or "MATTHEWS2016".}

\item{colland}{character. Defines the color of the land masses. By default colland = "#66666660".}

\item{colsea}{character. Defines the color of the sea. By default colsea = "#00509010".}

\item{colpoints}{character. Defines the color of the occurrence-points. By default colpoints = "#65432190".}

\item{rich.col}{character. Defines the color of the richness curve. By default rich.col = "#654321".}

\item{pch}{numeric. Point symbol for plotting the occurences. By default pch = 21.}

\item{do.plot}{logical. Defines if a plot is created or not. By default do.plot = TRUE.}

\item{save.as}{character. Defines the format the plots should be saved. "tiff", "pdf", "jpeg" or "png".}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to plot, 
such as main = "my own title" or main.col = "red".}
}
\value{
data.frame
}
\description{
Calculates the latitudinal diversity of taxa (species, genera, families, orders) and creates a plot of the continental masses with the fossil occurrences and the latitudinal diversity.
}
