% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/mapast_package.R
\docType{package}
\name{mapast}
\alias{mapast}
\alias{mapast}
\alias{mapast-package}
\alias{mapast-package}
\title{mapast}
\usage{
mapast(model = "SETON2012", data, map = NULL, do.plot = TRUE, save.as = NULL,
                         colland = "#66666660",
                         colsea = "#00509010", 
                         colpoints = "#65432190", 
                         pch = 16, cex = 1, ...)
}
\arguments{
\item{model}{character. Defining the model the map should be created with. "SETON2012" (default), 
"MULLER2016", "GOLONKA", "PALEOMAP" or "MATTHEWS2016".}

\item{data}{data.frame. Fossil occurrences data.}

\item{map}{(list of) SpatialPolygonDataFrames. Containing map(s) which can be created by getmap.}

\item{do.plot}{logical. Defines if a plot is created or not. By default do.plot = TRUE.}

\item{save.as}{character. Defines the format the plots should be saved. "tiff", "pdf", "jpeg" or "png".}

\item{colland}{character. Defines the color of the land masses. By default colland = "#66666660".}

\item{colsea}{character. Defines the color of the sea. By default colsea = "#00509010".}

\item{colpoints}{character. Defines the color of the occurrence-points. By default colpoints = "#65432190".}

\item{pch}{numeric. Point symbol for plotting the occurences. By default pch = 16 (filled circle).}

\item{cex}{numeric. Size of the points. By default cex = 1.}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to plot, such as main="my own title", main.col="red"}
}
\value{
Plot
}
\description{
Plots your fossil occurrence data onto the corresponding map.

We have developed mapast (PAST mAps), an R-package designed to make paleontolgical maps available to
plot the fossil occurrences from the Paleobiology Database onto paleogeographical maps.
}
\details{
We programmed functions for getting the SpatialPolygonsDaraFrame of the paleontological maps, 
functions to get a data.frame of fossil occurrences from the Paleobilogy Database, analysing the diversity of the past 
and plotting this data on the maps.

\tabular{ll}{
Package: \tab mapast\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2017-12-12\cr
License: \tab GPL-2\cr
}
}
\examples{
\dontrun{

library(mapast)

#get data and preprocess it
data  <-  base::data.frame(paleobioDB::pbdb_occurrences(base_name = "Canis", 
                                                        min_ma = 0, max_ma = 10, 
                                                        show = c("coords", "phylo"), 
                                                        vocab = "pbdb", limit = 100))
df <- formatdata(data)
df_auto <- paleocoords(df, time = "automatic")
                                                        
#create a plot with fossils on the paleogeographical map
mapast(model = "SETON2012", data = df_auto)

#save the maps before so the function does not need to load them
maps <- getmap(ma = 2.5, model = "SETON2012, do.plot = FALSE)
mapast(model = "SETON2012", data = df_auto, map = maps)

#save maps as pdf
mapast(model = "SETON2012", data = df_auto, map = maps, save.as = "pdf)


}
}
\seealso{
GPlates
\url{http://www.gplates.org/}

Golonka & Smith Global Paleoshorelines
\url{https://github.com/chhei/Heine_AJES_15_GlobalPaleoshorelines}

PaleobioDB
\url{http://paleobiodb.org}
}
\author{
Sonja Rothkugel \email{rothkugelsonja@aol.de}

Sara Varela \email{svarela@paleobiogeography.org}
}
\keyword{package}
