% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mapdiv}
\alias{mapdiv}
\title{mapdiv}
\usage{
mapdiv  (shape, data, unity, rank = "genus", res = 1, fun = mean,
                           colland = "#66666660", colsea = "#00509010", 
                           col.grid = mycols(100), do.plot = TRUE, ...)
}
\arguments{
\item{shape}{SpatialPolygonsDataFrame. A map which can be created with getmap.}

\item{data}{data.frame. Fossil occurrences data.}

\item{unity}{character. Either "fossilsite" or "cell".}

\item{rank}{character. Taxnomic rank. By default rank = "genus".}

\item{res}{numeric. Defining the spatial resolution. Default res = 1.}

\item{fun}{function or character. To determine what values to assign to cells that are covered by multiple spatial features. 
You can use functions such as min, max, or mean, or the character value: 'count'.}

\item{colland}{character. Defines the color of the land masses. By default colland = "#66666660".}

\item{colsea}{character. Defines the color of the sea. By default colsea = "#00509010".}

\item{col.grid}{character. Defines the color of the raster.}

\item{do.plot}{logical. Defines if a plot is created or not. By default do.plot = TRUE.}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to
plot, such as main = "my own title" or main.col = "red".}
}
\value{
RasterLayer
}
\description{
Calculates the Shannon diversity per cell or locality 
(taking into account relative abundances of all the fossil records 
whithin the cell) and creates a plot of the map with a RasterLayer of the diversity.
}
\examples{
\dontrun{

library(pasta)

#get GPlates map
shape <- getmap(interval = "Quaternary", model = "GPlates", do.plot = FALSE)
#get data from the paleobiology DB
data <- getdata (base_name = "Canis", interval = "Quaternary")
#calculate diversity per locality
div_site <- mapdiv (shape, data, unity = "fossilsite", rank ="genus", res = 1)
#calculate diversity per cell
div_cell <- mapdiv (shape, data, unity = "cell", rank = "genus", res = 1)

#save plot as pdf file
pdf("mapdiv-GPlates_Quaternary-Canis.pdf")
mapdiv (shape, data, unity = "fossilsite", rank = "genus", res = 1)
dev.off()

#save plot as tiff image
tiff("mapdiv-GPlates_Quaternary-Canis.tiff", 
      height = 10.5, width = 19, units = "cm", res = 300)
mapdiv (shape, data, unity = "fossilsite", rank = "genus", res = 1)
dev.off()

}
}
