% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mapdiv}
\alias{mapdiv}
\title{mapdiv}
\usage{
mapdiv  (data, unity, rank = "genus", res = 1, map=NULL, fun = mean, model="SETON2012",
                      colland = "#66666660", colsea = "#00509010", 
                      col.grid = mycols(100), 
                      do.plot = TRUE, save.as=NULL, ...)
}
\arguments{
\item{data}{data.frame. Fossil occurrences data.}

\item{unity}{character. Either "fossilsite" or "cell".}

\item{rank}{character. Taxnomic rank. By default rank = "genus".}

\item{res}{numeric. Defining the spatial resolution. Default res = 1.}

\item{map}{(list of) SpatialPolygonDataFrames. Containing map(s) which can be created by getmap.}

\item{fun}{function or character. To determine what values to assign to cells that are covered by multiple spatial features. 
You can use functions such as min, max, or mean, or the character value: 'count'.}

\item{model}{character. Defining the model the map should be created with. 'SETON2012' (default), 
'MULLER2016', 'GOLONKA', 'PALEOMAP' or 'MATTHEWS2016'.}

\item{colland}{character. Defines the color of the land masses. By default colland = "#66666660".}

\item{colsea}{character. Defines the color of the sea. By default colsea = "#00509010".}

\item{col.grid}{character. Defines the color of the raster.}

\item{do.plot}{logical. Defines if a plot is created or not. By default do.plot = TRUE.}

\item{save.as}{character. Defines the format the plots should be saved. "tiff", "pdf", "jpeg" or "png".}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to
plot, such as main = "my own title" or main.col = "red".}
}
\value{
RasterLayer
}
\description{
Calculates the Shannon diversity per cell or locality 
(taking into account relative abundances of all the fossil records 
whithin the cell) and creates a plot of the map with a RasterLayer of the diversity.
}
\examples{
\dontrun{

library(mapast)

#get GPlates map
shape <- getmap(interval = "Quaternary", model = "GPlates", do.plot = FALSE)
#get data from the paleobiology DB
data <- getdata (base_name = "Canis", interval = "Quaternary")
#calculate diversity per locality
div_site <- mapdiv (shape, data, unity = "fossilsite", rank ="genus", res = 1)
#calculate diversity per cell
div_cell <- mapdiv (shape, data, unity = "cell", rank = "genus", res = 1)

}
}
