% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mapocc}
\alias{mapocc}
\title{mapocc}
\usage{
mapocc(data, model = "SETON2012",
                   rank = "genus", map = NULL,
                   res = 1, save.as = NULL,
                   colland = "#66666660",
                   colsea = "#00509010", col.grid = mycols(100), do.plot = TRUE, ...)
}
\arguments{
\item{data}{data.frame. Fossil occurrences data.}

\item{model}{character. Defining the model the map should be created with. "SETON2012" (default), 
"MULLER2016", "GOLONKA", "PALEOMAP" or "MATTHEWS2016".}

\item{rank}{character. Defining the taxonomic rank of interest. 
"species", "genus", "family", "order", "class" or "phylum".}

\item{map}{(list of) SpatialPolygonDataFrames. Containing map(s) which can be created by getmap.}

\item{res}{numeric. Defining the spatial resolution. Default res = 1.}

\item{save.as}{character. Defines the format the plots should be saved. "tiff" or "pdf".}

\item{colland}{character. Defines the color of the land masses. By default colland = "#66666660".}

\item{colsea}{character. Defines the color of the sea. By default colsea = "#00509010".}

\item{col.grid}{character. Defines the color of the raster.}

\item{do.plot}{logical. Defines if a plot is created or not. By default do.plot = TRUE.}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to
plot, such as main = "my own title" or main.col = "red".}
}
\value{
RasterLayer
}
\description{
Creates a RasterLayer, containing the number of occurrences per cell, and a plot of the fossil 
occurences by taxonomic rank per cell (a proxy for the sampling effort).
}
