% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mapocc}
\alias{mapocc}
\title{mapocc}
\usage{
mapocc(shape, data, rank = "genus", res = 1,
                    colland = "#66666660", colsea = "#00509010", col.grid = mycols(100), 
                    do.plot = TRUE, ...)
}
\arguments{
\item{shape}{SpatialPolygonsDataFrame. A map that can be created with getmap.}

\item{data}{data.frame. Fossil occurrences data. Can be created with 
getdata(interval, base_name).}

\item{rank}{character. Defining the taxonomic rank of interest. 
"species", "genus", "family", "order", "class" or "phylum".}

\item{res}{numeric. Defining the spatial resolution. Default res = 1.}

\item{colland}{character. Defines the color of the land masses. By default colland = "#66666660".}

\item{colsea}{character. Defines the color of the sea. By default colsea = "#00509010".}

\item{col.grid}{character. Defines the color of the raster.}

\item{do.plot}{logical. Defines if a plot is created or not. By default do.plot = TRUE.}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to
plot, such as main = "my own title" or main.col = "red".}
}
\value{
RasterLayer
}
\description{
Creates a RasterLayer, containing the number of occurrences per cell, and a plot of the fossil 
occurences by taxonomic rank per cell (a proxy for the sampling effort).
}
\examples{
\dontrun{

library(pasta)

#get map
shape <- getmap(interval = "Quaternary", model = "GPlates", do.plot = FALSE)
#get data from paleobioDB
data <- getdata(base_name = "Canis", interval = "Quaternary")
#creasre map with occurrence RasterLayer
mapocc(shape, data)

#save plot as pdf file
pdf("mapocc-GPlates_Quaternary-Canis.pdf")
mapocc(shape, data)
dev.off()

# save plot as tiff image
tiff("mapocc-GPlates_Quaternary-Canis.tiff", 
      height = 10.5, width = 19, units = "cm", res = 300)
mapocc(shape, data)
dev.off()

}
}
