% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{maprich}
\alias{maprich}
\title{maprich}
\usage{
maprich(data, rank = "genus", res = 1, model = "SETON2012", map = NULL, save.as = NULL,
                    colland = "#66666660",
                    colsea = "#00509010", col.grid = mycols(100), do.plot = TRUE, ...)
}
\arguments{
\item{data}{data.frame. Fossil occurrences data. Can be created with 
getdata(interval, base_name).}

\item{rank}{character. Defining the taxonomic rank of interest. 
"species", "genus", "family", "order", "class" or "phylum". By default rank="genus".}

\item{res}{numeric. Defining the spatial resolution. Default res = 1.}

\item{model}{character. Defining the model the map should be created with. "SETON2012" (default), 
"MULLER2016", "GOLONKA", "PALEOMAP" or "MATTHEWS2016".}

\item{map}{(list of) SpatialPolygonDataFrames. Containing map(s) which can be created by getmap.}

\item{save.as}{character. Defines the format the plots should be saved. "tiff" or "pdf".}

\item{colland}{character. Defines the color of the land masses. By default colland = "#66666660".}

\item{colsea}{character. Defines the color of the sea. By default colsea = "#00509010".}

\item{col.grid}{character. Defines the color of the raster.}

\item{do.plot}{logical. Defines if a plot is created or not. By default do.plot=TRUE.}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to
plot, such as main = "my own title" or main.col = "red".}
}
\value{
RasterLayer
}
\description{
Creates a RasterLayer of taxon richness
and makes a plot of the map and richness raster.
}
\examples{
\dontrun{

library(mapast)

#get map
shape <- getmap(interval = "Paleocene", model = "GPlates")
#get the data from the paleobioDB
data<- getdata(base_name = "Testudines", interval = "Paleocene")
#calculate the genus richness
richness<- maprich(shape, data, rank = "genus")

}

}
