% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{maprich}
\alias{maprich}
\title{maprich}
\usage{
maprich(data, rank = "genus", res = 1, model = "SETON2012", map = NULL, save.as = NULL,
                    colland = "#66666660",
                    colsea = "#00509010", col.grid = mycols(100), do.plot = TRUE, ...)
}
\arguments{
\item{data}{data.frame. Fossil occurrences data.}

\item{rank}{character. The taxonomic rank of interest. 
"species", "genus", "family", "order", "class" or "phylum". By default rank="genus".}

\item{res}{numeric. The spatial resolution. By default res = 1.}

\item{model}{character. Defining the model the map should be created with. "SETON2012" (default), 
"MULLER2016", "GOLONKA", "PALEOMAP" or "MATTHEWS2016".}

\item{map}{(list of) SpatialPolygonDataFrames. Containing map(s) which can be created by mapast::getmap(ma, model).}

\item{save.as}{character. The format the plots should be saved. "tiff", "pdf", "jpeg" or "png".}

\item{colland}{character. The color of the land masses. By default colland = "#66666660".}

\item{colsea}{character. The color of the sea. By default colsea = "#00509010".}

\item{col.grid}{character. The color of the raster. By default col.grid = mycols(100).}

\item{do.plot}{logical. If a plot is created or not. By default do.plot = TRUE.}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to
plot, such as main = "my own title" or main.col = "red".}
}
\value{
RasterLayer
}
\description{
Creates a RasterLayer of taxon richness
and makes a plot of the map with the richness raster.
}
\examples{
\dontrun{

library(mapast)

#get data and preprocess it
data  <-  base::data.frame(paleobioDB::pbdb_occurrences(base_name = "Canis", 
                                                        min_ma = 0, max_ma = 10, 
                                                        show = c("coords", "phylo"), 
                                                        vocab = "pbdb", limit = 100))
df <- formatdata(data = data)
df_auto <- paleocoords(data = df, time = "automatic")
                                                        
#create a plot with fossils on the paleogeographical map
rich <- maprich(data = df_auto, rank = "species", res = 10, model = "SETON2012")

#save the maps before so the function does not need to load them
maps <- getmap(ma = 2.5, model = "SETON2012", do.plot = FALSE)
maprich(data = df_auto, rank = "species", res = 10, model = "SETON2012", map = maps)

#save maps as pdf
maprich(data = df_auto, rank = "species", res = 10, model = "SETON2012", map = maps,
                      save.as = "pdf")

}
}
