% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{paleocoords}
\alias{paleocoords}
\title{paleocoords}
\usage{
paleocoords(data, time = "automatic", timevector = NULL, 
                        stepsize = 10, model = "SETON2012")
}
\arguments{
\item{data}{data.frame. Fossil occurrences data.}

\item{time}{character. Defines how the reconstruction time is specified. Can be "automatic", "average" or "timevector". By default time="automatic".}

\item{timevector}{vector. Defining the borders of the time bins. Not allowed to be NULL if time="timevector".}

\item{stepsize}{numeric. Defining the stepsize of the time bins if time="automatic".}

\item{model}{character. Defining the model the map should be created with. "SETON2012" (default), 
"MULLER2016", "GOLONKA", "PALEOMAP" or "MATTHEWS2016".}
}
\value{
data.frame
}
\description{
Calculating the paleocoordinates of the fossil data. Either calculating timebins by looking at the early_age and late_age column,
taking the average age of the fossils rounded or by using user defined time bins.
}
\examples{
\dontrun{

occ <- base::data.frame(paleobioDB::pbdb_occurrences(base_name="Canis", min_ma = 0, max_ma = 10, 
                                                     show = c("coords", "phylo"), 
                                                     vocab = "pbdb", limit = 100))

#reconstruct paleocoordinates with midpoint of appearance time
occ_ma <- paleoocoords(occ, time="automatic", model = "SETON2012")

#reconstruct paleocoordinates with specific time
occ_matime <- paleoocoords(occ, time ="timevector", timevector = c(0.10), model = "SETON2012")

                    
}
}
