% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleoMap_package.R
\docType{package}
\name{paleoMap}
\alias{paleoMap}
\alias{paleoMap-package}
\alias{paleoMap-package}
\title{paleoMap: An R-package for getting and using paleontological maps.}
\description{
We have developed paleoMap, an R-package designed to make palaeontolgical maps 
available to plot the fossil occurences from paleontological databases onto the maps of the past.
}
\details{
We programmed functions for getting the shape files of the paleo-maps, 
functions to get fossil data from the paleobilogy database and plotting this 
data on the maps.

\tabular{ll}{
Package: \tab paleoMap\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2016-01-01\cr
License: \tab GPL-2\cr
}
}
\seealso{
GPlates
\url{http://www.gplates.org/}

PaleobioDB
\url{http://paleobiodb.org}
}
\author{
Sonja Rothkugel \email{rothkugelsonja@aol.de}

Sara Varela \email{svarela@paleobiogeography.org}
}
\keyword{package}
