% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_divraster_cell}
\alias{pm_divraster_cell}
\title{pm_divraster_cell}
\usage{
pm_divraster_cell  (shape, occ_df_cell, res=10, rank="species",
                           colsea="#00509010", colland="#66666680")
}
\arguments{
\item{shape}{file from the time interval of interest. 
Can be created with get_paleomap}

\item{occ_df_cell}{a data frame with number of occurrences of 
a taxa per locality}

\item{res}{resolution of the raster/ size of the grid cell}

\item{rank}{taxonomic rangk of interest. values: "species", "genus", 
"family", "order". By default rank="species"}

\item{colsea}{color of the ocean}

\item{colland}{color of the land masses}
}
\value{
plot with map of the time intervall, the fossil occurences and the 
raster file. And the raster file itself
}
\description{
calculates the Shannon diversity per cell 
(taking into account relative abundances of all the fossil records 
whithin the cell)
}
\examples{
\dontrun{
shape<- pm_getmap(interval="Quaternary") 
data<- pm_getdata (base_name="Canis", interval="Quaternary")
occ_df_cell <- pm_occ_cell (data, rank="species")
pm_divraster_cell (shape, occ_df_cell, res=10)
## cells with diversity values = 0 (e.g., 1 species) are discarded.
}
}
