% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_divraster_loc}
\alias{pm_divraster_loc}
\title{pm_divraster_loc}
\usage{
pm_divraster_loc  (shape, occ_df, res=10, fun=mean,
                          colsea="#00509010", colland="#66666680", ...)
}
\arguments{
\item{shape}{file from the time interval of interest. 
Can be created with get_paleomap}

\item{occ_df}{a data frame with number of occurrences of a taxa per locality}

\item{res}{resolution of the raster/ size of the grid cell}

\item{fun}{values: mean, max, min, "count". functions to be applied when 
making the raster use mean to get the mean value of diversity in the cell
(mean of the different fossil sites), max to get the maximum diversity, 
min to get the min value of diversity, "count" to get the number of 
fossil sites in per cell}

\item{colsea}{color of the ocean}

\item{colland}{color of the land masses}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to plot, such as main="my own title", main.col="red"}
}
\value{
plot with map of the time intervall, the fossil occurences and the 
raster file. And the raster file itself
}
\description{
calculates the Shannon diversity per unique locality (based on 
its coordinates), makes a raster file and a plot showing mean, 
max, min diversity per cell, or number of unique localities per cell
}
\examples{
\dontrun{
shape<- pm_getmap(interval="Quaternary", model="GPlates", do.plot=FALSE)
data<- pm_getdata (base_name="Canis", interval="Quaternary")
occ_df <- pm_occ (data, rank="species")
pm_divraster_loc (shape, occ_df, fun=mean)
pm_divraster_loc (shape, occ_df, fun=max)
pm_divraster_loc (shape, occ_df, fun=min)
pm_divraster_loc (shape, occ_df, fun="count")
}
}
