% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_getdata}
\alias{pm_getdata}
\title{pm_getdata}
\usage{
pm_getdata(interval, base_name, limit="all")
}
\arguments{
\item{interval}{time interval of interest (e.g. jurassic)}

\item{base_name}{name of the taxon you want to get data from 
(e.g mammalia, canidae, canis)}

\item{limit}{numeric. To set the limit of occurrences 
to be downloaded from the Paleobiology Database e.g. 500. 
There is no limit by default.}
}
\value{
a data frame with the occurrences and the values needed 
for the other functions in paleoMap
}
\description{
uses paleobioDB R package to get data from the Paleobiology Database
}
\examples{
\dontrun{
pm_getdata(interval="Quaternary",base_name="Canis")

#if you want to use FROMAGE and TOAGE to define the interval you can use the paleobioDB package
myocc <- base::data.frame(paleobioDB::pbdb_occurrences(base_name="mammalia", 
                    min_ma=0, max_ma=2.58, 
                    show=c("paleoloc", "phylo"), 
                    vocab="pbdb", limit=100))
}
}
