% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_getdata}
\alias{pm_getdata}
\title{pm_getdata}
\usage{
pm_getdata (interval, base_name, limit)
}
\arguments{
\item{interval}{time interval of interest (e.g. jurassic)}

\item{base_name}{name of the base (e.g reptilia, mammalia, canis) you want to get data from}

\item{limit}{how many entrances from pbdb you want to have, e.g. 500.
There is no limit by default}
}
\value{
a data frame with the occurrences and the values needed for the other functions in paleoMap
}
\description{
uses paleobioDB R package to get data from paleobiology Database
}
\examples{
\dontrun{
pm_getdata (base_name="Canis", interval="Quaternary")
}
}

