% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_getmap}
\alias{pm_getmap}
\title{pm_getmap}
\usage{
pm_getmap(interval, model, colland = "#66666660"
                          , colsea = "#00509010", do.plot = TRUE, ...)
}
\arguments{
\item{interval}{character. Temporal paleoeographical interval of interest. (e.g. "Cretaceous" for GPlates or "112.0" for Smith)}

\item{model}{character. Defining the model the map was created with. "GPlates", "Smith" or "Golonka"}

\item{colland}{define the color of the land masses. By default colland = "#66666660"}

\item{colsea}{define the color of the sea. By default colsea = "#00509010"}

\item{do.plot}{logical. Defines if a plot is created or not. By default do.plot=TRUE}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to plot, such as main="my own title", main.col="red"}
}
\value{
a SpatialPolygonsDataFrame and a plot (if do.plot=TRUE)
}
\description{
generates a shapefile with the paleomap of the chosen 
time interval (e.g. "Cretaceous") and a plot
}
\examples{
\dontrun{
pm_getmap(interval="Cretaceous", model="GPlates")

#for checking which maps are available please use
d <- data(package='paleogeoDB')
d$results[, "Item"]
}
}
