% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_getmap}
\alias{pm_getmap}
\title{pm_getmap}
\usage{
pm_getmap (interval, do.plot, colsea, colland, colborder)
}
\arguments{
\item{interval}{time interval of interest}

\item{do.plot}{TRUE/FALSE, if TRUE the output includes a plot}

\item{colsea}{to set the color of the ocean in the plot}

\item{colland}{to set the color of the land masses}

\item{colborder}{to set the color of the borders of the land masses}
}
\value{
a shape file for the choosen time interval and a plot (if do.plot=TRUE)
}
\description{
gets the shapefile for the map of a choosen time interval (e.g. "triassic")
}
\examples{
/dontrun{
pm_getmap(interval="Quaternary")
}
}

