% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_latdiv}
\alias{pm_latdiv}
\title{pm_latdiv}
\usage{
pm_latdiv (shape, occ_df, res=10, fun= max, magn=1,
                  colland="#66666680", colsea="#00509010",  
                  colpoints="#FFC12530",colpointborder="black", 
                  div.col="goldenrod1", pch=21, ...)
}
\arguments{
\item{shape}{SpatialPolygonsDataFrame object containing a map.}

\item{occ_df}{data.frame with fossil occurrences. Can be created with pm_occ(data)}

\item{res}{numeric. Defining the spatial resolution. By default res=10.}

\item{fun}{function or character. To determine what values to assign to cells that are covered by multiple spatial features. 
You can use functions such as min, max, or mean, or the character value: 'count'.}

\item{magn}{numeric. Defining the magnitude of the diversity curve.}

\item{colland}{define the color of the land masses. By default colland = "#66666660".}

\item{colsea}{define the color of the sea. By default colsea = "#00509010".}

\item{colpoints}{define the color of the occurrence-points. By default colpoints="#FFC12530".}

\item{colpointborder}{define the color of the occurrence-points border. Bx default colpointborder="black".}

\item{div.col}{define the color od zje diversity curve. By default div.col="goldenrod1".}

\item{pch}{point symbol for plotting the occurences. By default pch=16 (filled circle).}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to plot, such as main="my own title", main.col="red"}
}
\value{
data frame with shannon diversity,
a plot of the corresponding time map with the occurrences and their
latitudinal diversity
}
\description{
calculates the Shannon diversity along the latitudinal gradient based on 
the individual values of diverstiy the fossil localities of those latitudes.
The function returns the mean or max values of diversity of the sampled 
localities along the latitudinal gradient.
}
\examples{
\dontrun{
shape<- pm_getmap(interval="Quaternary", model="GPlates", do.plot=FALSE)
data<- pm_getdata (base_name="Canis", interval="Quaternary")
occ_df <- pm_occ (data)
pm_latdiv (shape, occ_df, fun=mean, magn=10)
pm_latdiv (shape, occ_df, fun=max, magn=5)
}
}
