% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_latdiv}
\alias{pm_latdiv}
\title{pm_latdiv}
\usage{
pm_latdiv (shape, occ_df, res=10, fun= max, magn=1,
                  colsea="#00509010", colland="#66666680", 
                  colpoints="#FFC12530",colpointborder="black", div.col="goldenrod1"
                  , pch=21, ...)
}
\arguments{
\item{shape}{a shape file of the corresponding time map}

\item{occ_df}{a data frame with abundance of taxa per locality (see example)}

\item{res}{numeric. spatial resolucion of the latitudinal bins. 
res=10 by default.}

\item{fun}{values: mean, max. functions to be applied to get one single value 
of diversity, use mean to get the mean value of diversity 
(mean of the different fossil sites at the same latitudinal bin), 
max to get the maximum observed diversity, min to get the mean 
observed diversity.}

\item{magn}{numeric. index to magnify the plot of the latitudinal richness 
(use when richness values are very low to see the latitudinal pattern more easily)}

\item{colsea}{defines the color of the sea}

\item{colland}{defines the color f the landmasses}

\item{colpoints}{defines the colo of the points for the occurrences}

\item{colpointborder}{defines color of the border of the occurrence points}

\item{div.col}{color for richnness graph}

\item{pch}{point symbol for plotting the occurences. By default pch=16 (filled circle)}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to plot, such as main="my own title", main.col="red"}
}
\value{
data frame with shannon diversity,
a plot of the corresponding time map with the occurrences and their
latitudinal diversity
}
\description{
calculates the Shannon diversity along the latitudinal gradient based on 
the individual values of diverstiy the fossil localities of those latitudes.
The function returns the mean or max values of diversity of the sampled 
localities along the latitudinal gradient.
}
\examples{
\dontrun{
shape<- pm_getmap(interval="Quaternary", model="GPlates", do.plot=FALSE)
data<- pm_getdata (base_name="Canis", interval="Quaternary")
occ_df <- pm_occ (data)
pm_latdiv (shape, occ_df, fun=mean, magn=10)
pm_latdiv (shape, occ_df, fun=max, magn=5)
}
}
