% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_latdiv}
\alias{pm_latdiv}
\title{pm_latdiv}
\usage{
pm_latdiv (occ_df, shape, fun, colsea, solland, colpoints, colpointborder)
}
\arguments{
\item{occ_df}{a data frame with abundance of taxa per locality (see example)}

\item{shape}{a shape file of the corresponding time map}

\item{fun}{values: mean, max. functions to be applied to get one single value of diversity,
use mean to get the mean value of diversity 
(mean of the different fossil sites at the same latitudinal bin), 
max to get the maximum observed diversity, min to get the min observed diversity.}

\item{colsea}{defines the color of the sea}

\item{colland}{defines the color f the landmasses}

\item{colpoints}{defines the colo of the points for the occurrences}

\item{colpointborder}{defines color of the border of the occurrence points}

\item{res.}{numeric. spatial resolucion of the latitudinal bins. res=10 by default.}
}
\value{
data frame with shannon diversity,
a plot of the corresponding time map with the occurrences and their
latitudinal diversity
}
\description{
calculates the Shannon diversity along the latitudinal gradient based on 
the individual values of diverstiy the fossil localities of those latitudes.
The function returns the mean or max values of diversity of the sampled localities
along the latitudinal gradient.
}
\examples{
\dontrun{
data<- pm_getdata (base_name="Canis", interval="Quaternary")
occ_df <- pm_occ (data)
pm_latdiv (occ_df, shape, fun=mean)
pm_latdiv (occ_df, shape, fun=max)
}
}
