% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_latrich}
\alias{pm_latrich}
\title{pm_latrich}
\usage{
pm_latrich (data, res)
}
\arguments{
\item{data}{a data frame with fossil occurrences
Can be created with pm_getdata(interval, base_name)}

\item{res}{resolution in of the segmentation of the latitude}
}
\value{
data frame with richness of rank
}
\description{
calculates the latitudinal richness of the genus
}
\examples{
/dontrun{
data<- pm_getdata (base_name="Canis", interval="Quaternary")
latrich <- pm_latrich (data, res=10)
show(latrich)
}
}

