% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_latrich}
\alias{pm_latrich}
\title{pm_latrich}
\usage{
pm_latrich (shape, data, rank="species", res=10,
                   colland="#66666680", colsea="#00509010", 
                   colpoints="#8B696980",colpointborder="#8B696980", 
                   rich.col="#8B696980", pch=21, do.plot=TRUE, ...)
}
\arguments{
\item{shape}{SpatialPolygonsDataFrame object containing a map.}

\item{data}{data.frame with fossil occurrences. Can be created with
pm_getdata(interval, base_name)}

\item{rank}{character. Defining the taxonomic rank of interest. 
"species", "genus", "family", "order", "class" or "phylum". By default rank="species".}

\item{res}{numeric. Defining the spatial resolution. Default res=10.}

\item{colland}{define the color of the land masses. By default colland = "#66666660".}

\item{colsea}{define the color of the sea. By default colsea = "#00509010".}

\item{colpoints}{define the color of the occurrence-points. By default colpoints="#8B696980".}

\item{colpointborder}{define the color of the occurrence-points border. By default colpointborder="#8B696980".}

\item{rich.col}{define the color of the richness curve. By default rich.col="#8B696980".}

\item{pch}{point symbol for plotting the occurences. By default pch=21.}

\item{do.plot}{logical. Defines if a plot is created or not. By default do.plot=TRUE.}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to plot, 
such as main="my own title", main.col="red".}
}
\value{
data frame with richness of rank and a plot of the continental masses 
with the occurrences and the latitudinal richness
}
\description{
calculates latitudinal diversity of taxa (species, genera, families, orders)
}
\examples{
\dontrun{
shape<- pm_getmap(interval="Quaternary", model="GPlates", do.plot=FALSE)
data<- pm_getdata (base_name="Canis", interval="Quaternary")
pm_latrich (shape, data, rank="species", res=10)
}
}
