% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_occ}
\alias{pm_occ}
\title{pm_occ}
\usage{
pm_occ (data, rank)
}
\arguments{
\item{data}{a data frame with fossil occurrences 
Can be created with pm_getdata(interval, base_name)}

\item{rank}{character: "species", "genus", "family", "order". 
By default rank="species"}
}
\value{
data frame with number of species, genera, families or orders per locality
}
\description{
generates a diversity matrix, with the number occurrences of each species, genus, family or order per locality
}
\examples{
\dontrun{
data<- pm_getdata (base_name="Canis", interval="Quaternary")
pm_occ (data, rank="species")
}
}

