% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_occ_cell}
\alias{pm_occ_cell}
\title{pm_occ_cell}
\usage{
pm_occ_cell(data, rank = "species", res = 10)
}
\arguments{
\item{data}{a data frame with fossil occurrences 
Can be created with pm_getdata(interval, base_name)}

\item{rank}{character: "species", "genus", "family", "order". 
By default rank="species"}

\item{res}{numeric. resolution of the cells. By default res=10}

\item{shape}{file from the time interval of interest. 
Can be created with get_paleomap}
}
\value{
data frame with number of occurrences by species, genera, families or orders per locality
}
\description{
generates a diversity matrix, with the number occurrences of each species, genus, family or order per cell
}
\examples{
\dontrun{
data <- pm_getdata(base_name = "Canis", interval = "Quaternary")
result <- pm_occ_cell(data, rank = "species", res = 10)
}
}
