% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_occraster}
\alias{pm_occraster}
\title{pm_occraster}
\usage{
pm_occraster(shape, data, rank = "species", res = 10,
                    colsea = "#00509010", colland = "#66666660", ...)
}
\arguments{
\item{shape}{shapefile from the time interval of interest. 
It can be created with pm_getmap}

\item{data}{a data frame which needs to have a column called paleolat 
and a column called paleolng. It can be created with getdata_paleomap}

\item{rank}{taxonomic rank of interest (e.g. genus, family, etc.). 
By default rank="species"}

\item{res}{resolution of the cells in the raster in degrees (by default res=10)}

\item{colsea}{users can define the color of the ocean}

\item{colland}{users can define the color of the land masses}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to plot, such as main="my own title", main.col="red"}
}
\value{
a raster file and a plot with number of the fossil 
occurrences in the selected interval at the selected resolution
}
\description{
creates a raster and a plot of the fossil occurences by taxonomic rank per cell 
(a proxy for the sampling effort)
}
\examples{
\dontrun{
shape <- pm_getmap(interval="Quaternary", model="GPlates", do.plot = FALSE)
data <- pm_getdata(base_name="Canis", interval="Quaternary")
pm_occraster(shape, data)
}
}
