% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_occraster}
\alias{pm_occraster}
\title{pm_occraster}
\usage{
pm_occraster(shape, data, rank = "species", res = 10,
                    colland = "#66666660", colsea = "#00509010", col.grid=mycols(100), 
                    do.plot=TRUE, ...)
}
\arguments{
\item{shape}{SpatialPolygonsDataFrame object containing a map.}

\item{data}{data.frame with fossil occurrences. Can be created with 
pm_getdata(interval, base_name)}

\item{rank}{character. Defining the taxonomic rank of interest. 
"species", "genus", "family", "order", "class" or "phylum".}

\item{res}{numeric. Defining the spatial resolution. Default res=10.}

\item{colland}{define the color of the land masses. By default colland = "#66666660".}

\item{colsea}{define the color of the sea. By default colsea = "#00509010".}

\item{col.grid}{define the color of the raster.}

\item{do.plot}{logical. Defines if a plot is created or not. By default do.plot=TRUE.}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to
plot, such as main="my own title", main.col="red".}
}
\description{
creates a raster and a plot of the fossil occurences by taxonomic rank per cell 
(a proxy for the sampling effort)
}
\examples{
\dontrun{
shape <- pm_getmap(interval="Quaternary", model="GPlates", do.plot = FALSE)
data <- pm_getdata(base_name="Canis", interval="Quaternary")
pm_occraster(shape, data)
}
}
