% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_occraster}
\alias{pm_occraster}
\title{pm_occraster}
\usage{
pm_occraster (shape, data, rank, res, colsea, colland, colborder)
}
\arguments{
\item{shape}{shapefile from the time interval of interest.
Can be created with pm_getmap}

\item{data}{a data frame which needs to have a column called paleolat
and a column called paleolng, can be created with getdata_paleomap}

\item{rank}{Taxonomic rank of interest (e.g. genus, family, etc.)}

\item{res}{resolution of the cells in the raster (10 degrees by default)}

\item{colsea}{color of the ocean}

\item{colland}{color of the land masses}

\item{colborder}{color of the landmass borders}
}
\value{
a raster file and a plot of the time intervall, the fossil occurences and the raster file.
}
\description{
creates a raster and a plot of the fossil occurences by taxonomic rank per cell
(a proxy for the sampling effort)
}
\examples{
/dontrun{
shape<- pm_getmap(interval="Quaternary")
data<- pm_getdata (base_name="Canis", interval="Quaternary")
pm_occraster (shape, data)
}
}

