% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_plot}
\alias{pm_plot}
\title{pm_plot}
\usage{
pm_plot(interval, model, data,colsea = "#00509010",
                colland = "#66666660",colpoints = "#99000020", 
                cex = 1, pch =16, ...)
}
\arguments{
\item{interval}{time interval of interest (e.g. Quaternary)}

\item{model}{which reconstruction model the map comes from. 
"GPlates", "Golonka" or "Smith"}

\item{data}{data.frame with the paleogeoreferenced fossil records 
(can be obtained using pm_getdata)}

\item{colsea}{color of the ocean}

\item{colland}{color of the land masses}

\item{colpoints}{color of the occurrence points}

\item{cex}{numeric. size of the points. By default cex=1.}

\item{pch}{point symbol for plotting the occurences. By default pch=16 (filled circle)}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to plot, such as main="my own title", main.col="red"}
}
\value{
a plot with the configuration of the continents at the selected 
time interval and the fossil occurrences
}
\description{
plots your query from the paleobioDB onto the map of the selected time interval
}
\examples{
\dontrun{
data  <-  pm_getdata (base_name="Mammalia", interval="Cretaceous")
pm_plot(interval="Cretaceous", model="GPlates", data)
}
}
