% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_plot}
\alias{pm_plot}
\title{pm_plot}
\usage{
pm_plot (interval, base_name, limit, colsea, colland,
colborder, colpoints, colpointborder)
}
\arguments{
\item{interval}{time interval of interest (e.g. jurassic)}

\item{base_name}{larger taxonomic rank for the query to the paleobioDB (e.g reptiles)}

\item{limit}{enables the user to set the maximum number of
records downloaded from the paleobioDB (e.g. 1000)}

\item{colsea}{color of the ocean}

\item{colland}{color of the land masses}

\item{colborder}{color of the landmass borders}

\item{colpoints}{color of the points of the occurences}

\item{colpointborder}{color of the border of the points}
}
\value{
a plot with the configuration of the continents at the selected time interval
and the fossil occurrences
}
\description{
plots your query from paleobioDB directly onto the map of the selected time interval
}
\examples{
\dontrun{
pm_plot (interval="Quaternary", base_name= "Canis")
}
}

