% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_plot}
\alias{pm_plot}
\title{pm_plot}
\usage{
pm_plot(interval, model, data, colland = "#66666660",
                colsea = "#00509010",colpoints = "#65432190", 
                pch =16, cex = 1, ...)
}
\arguments{
\item{interval}{character. Temporal paleoeographical interval of interest. (e.g. "Cretaceous" for GPlates or "112.0" for Smith)}

\item{model}{character. Defining the model the map was created with. "GPlates", "Smith" or "Golonka"}

\item{data}{data.frame with fossil occurrences. Can be created with pm_getdata(interval, base_name)}

\item{colland}{define the color of the land masses. By default colland = "#66666660".}

\item{colsea}{define the color of the sea. By default colsea = "#00509010".}

\item{colpoints}{define the color of the occurrence-points. By default colpoints = "#65432190".}

\item{pch}{point symbol for plotting the occurences. By default pch=16 (filled circle).}

\item{cex}{numeric. size of the points. By default cex=1.}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to plot, such as main="my own title", main.col="red"}
}
\value{
a plot with the configuration of the continents at the selected 
time interval and the fossil occurrences
}
\description{
plots your query from the paleobioDB onto the map of the selected time interval
}
\examples{
\dontrun{
data  <-  pm_getdata (interval="Cretaceous", base_name="Mammalia")
pm_plot(interval="Cretaceous", model="GPlates", data)

#save as pdf file
pdf("pm_plot-GPlates_Cretaceous-Mammalia.pdf")
pm_plot(interval="Cretaceous", model="GPlates", data)
dev.off()
#save as tiff image
tiff("pm_plot-GPlates_Cretaceous-Mammalia.tiff", 
      height = 10.5, width = 17, units = 'cm', res=300)
pm_plot(interval="Cretaceous", model="GPlates", data)
dev.off()

}
}
