% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_richraster}
\alias{pm_richraster}
\title{pm_richraster}
\usage{
pm_richraster(shape, data, rank, res = 10, 
                     colland = "#66666660", colsea = "#00509010", col.grid= mycols(100), 
                     do.plot=TRUE, ...)
}
\arguments{
\item{shape}{SpatialPolygonsDataFrame object containing a map.}

\item{data}{data.frame with fossil occurrences. Can be created with 
pm_getdata(interval, base_name)}

\item{rank}{character. Defining the taxonomic rank of interest. 
"species", "genus", "family", "order", "class" or "phylum".}

\item{res}{numeric. Defining the spatial resolution. Default res=10.}

\item{colland}{define the color of the land masses. By default colland = "#66666660".}

\item{colsea}{define the color of the sea. By default colsea = "#00509010".}

\item{col.grid}{define the color of the raster.}

\item{do.plot}{logical. Defines if a plot is created or not. By default do.plot=TRUE.}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to
plot, such as main="my own title", main.col="red".}
}
\value{
Creates a plot with map of the time intervall, the fossil occurences and the 
RasterLayer. And returns the RasterLayer itself
}
\description{
Creates a raster of species richness
and makes a plot of the map and raster.
}
\examples{
\dontrun{

shape<- pm_getmap(interval="Paleocene", model="GPlates")
data<- pm_getdata(base_name="Testudines", interval="Paleocene")
richness<- pm_richraster(shape, data, rank="genus")

#save as pdf file
pdf("pm_richraster-GPlates_Paleocene-Testudines.pdf")
richness<- pm_richraster(shape, data, rank="genus")
dev.off()
#save as tiff image
tiff("pm_richraster-GPlates_Paleocene-Testudines.tiff", 
      height = 10.5, width = 19, units = 'cm', res=300)
richness<- pm_richraster(shape, data, rank="genus")
dev.off()

}

}
