% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pm_richraster}
\alias{pm_richraster}
\title{pm_richraster}
\usage{
pm_richraster(shape, data, res = 10, rank, 
                     colsea = "#00509010", colland = "#66666660", ...)
}
\arguments{
\item{shape}{file from the time interval of interest. 
Can be created with get_paleomap}

\item{data}{a data frame which needs to have a column called 
paleolat and a column called paleolng, can be created with getdata_paleomap}

\item{res}{resolution of the raster/ size of the grid cell}

\item{rank}{gives the rank for the richness raster (e.g. species, genus,...)}

\item{colsea}{color of the ocean}

\item{colland}{color of the land masses}

\item{...}{Graphical parameters. Any argument that can be passed to image.plot and to plot, such as main="my own title", main.col="red"}
}
\value{
plot with map of the time intervall, the fossil occurences and the 
raster file. And the raster file itself
}
\description{
Creates a raster of species richness
and makes a plot of the map and raster
}
\examples{
\dontrun{
shape<- pm_getmap(interval="Paleocene", model="GPlates")
data<- pm_getdata(base_name="Testudines", interval="Paleocene")
richness<- pm_richraster(shape, data, rank="genus")
}

}
