% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{rank_filter}
\alias{rank_filter}
\title{rank_filter}
\usage{
rank_filter (res, data, rank)
}
\arguments{
\item{data}{a data frame which needs to have a column called paleolat and a column called paleolng,
can be created with getdata_paleomap}

\item{res}{resolution of the raster file}

\item{rank}{rank of interest}

\item{ras}{blank raster}
}
\value{
a raster with the taxa richness
}
\description{
creates the raster files with the number of unique taxa by pixel
}
\examples{
\dontrun{
data<- pm_getdata (base_name="Canis", interval="Quaternary")
rank_filter (data, res=10, rank="genus")
}
}

