% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{spsite}
\alias{spsite}
\title{spsite}
\usage{
spsite(data, unity, res = 1, rank = "genus", pa = FALSE)
}
\arguments{
\item{data}{data.frame. Fossil occurrences data.}

\item{unity}{character. unity = "fossilsite" or unity = "cell" defining if the user wants the occurrences per 
cell or per fossilsite.}

\item{res}{numeric. Defining the spatial resolution. By default res = 1. Only used if unity = "cell".}

\item{rank}{character. Defining the taxonomic rank of interest. 
"species", "genus", "family", "order", "class" or "phylum". By default rank = "genus"}

\item{pa}{boolean. Defines if the user wants presence absence or counted data. By default pa = FALSE.}
}
\value{
data.frame
}
\description{
Generates a diversity data.frame, with the number occurrences of a taxon per locality.
}
\examples{
\dontrun{

library(mapast)

#get data and preprocess it
data  <-  base::data.frame(paleobioDB::pbdb_occurrences(base_name = "Canis", 
                                                        min_ma = 0, max_ma = 10, 
                                                        show = c("coords", "phylo"), 
                                                        vocab = "pbdb", limit = 100))
df <- formatdata(data)
df_auto <- paleocoords(df, time = "automatic")
                                                        
#create a plot with fossils on the paleogeographical map
sp_fossilsite <- spsite(data = df_auto, unity = "fossilsite", rank = "species", res = 10)
sp_cell <- spsite(data = df_auto, unity = "cell", rank = "species", res = 10)


}
}
